+++ src/app/main/DatabaseSolver.js
@@ -2,6 +2,9 @@ import * as DatabaseSetup from '../../database/DatabaseSetup.js';
 import * as ClientHandler from '../client/ClientHandler.js';
 import * as OutputHandler from '../main/OutputHandler.js';
 import * as FileUtil from '../../util/FileUtil.js';
+import * as DateUtil from '../../util/DateUtil.js';
+
+import * as ReviewDatabase from '../../database/ReviewDatabase.js';
 
 import * as ErrorReviewer from './ErrorReviewer.js';
 import * as Menu from '../client/menu/Menu.js';
@@ -63,7 +66,13 @@ export async function shouldSaveNewReviewsForClient(db, config, reviews)
     return result;
 }
 
-export async function outputNewReviewsToFile(db, config, reviews)
+/**
+ * Writes the reviews to file. If no reviews are passed-in, it will output the entire reviews database.
+ * @param {Database} db The database.
+ * @param {Config} config The config.
+ * @param {Array<Array>} [reviews] An array of new reviews.
+ */
+export async function outputNewReviewsToFile(db, config, reviews = null)
 {
     const path = require('path');
     
@@ -79,23 +88,48 @@ export async function outputNewReviewsToFile(db, config, reviews)
         '...'
     ];
     const reviewTable = [reviewTableHeader];
-    for(const review of reviews)
+
+    if (reviews)
     {
-        const reviewEntry = [];
-        // ID
-        reviewEntry.push(review[0]);
-        // Date
-        reviewEntry.push(review[1]);
-        // Comment
-        reviewEntry.push(review[2]);
-        // Type
-        reviewEntry.push(review[3]);
-        // Params
-        reviewEntry.push(...review[4]);
-        reviewTable.push(reviewEntry);
+        // Append ALL reviews (including new ones)
+        for(const reviewID of ReviewDatabase.getReviews(db))
+        {
+            const review = ReviewDatabase.getReviewByID(db, reviewID);
+            const reviewEntry = [];
+            // ID
+            reviewEntry.push(reviewID);
+            // Date
+            reviewEntry.push(DateUtil.stringify(review.date));
+            // Comment
+            reviewEntry.push(review.comment);
+            // Type
+            reviewEntry.push(review.type);
+            // Params
+            reviewEntry.push(...review.params);
+            reviewTable.push(reviewEntry);
+        }
+    }
+    else
+    {
+        // Append only NEW reviews
+        for(const review of reviews)
+        {
+            const reviewEntry = [];
+            // ID
+            reviewEntry.push(review[0]);
+            // Date
+            reviewEntry.push(review[1]);
+            // Comment
+            reviewEntry.push(review[2]);
+            // Type
+            reviewEntry.push(review[3]);
+            // Params
+            reviewEntry.push(...review[4]);
+            reviewTable.push(reviewEntry);
+        }
     }
 
-    const outputFilePath = path.resolve(config.outputPath, `reviews-${db.currentDate.toISOString()}.csv`);
+    const outputFilePath = path.resolve(config.outputPath, `reviews-${DateUtil.stringify(new Date(Date.now()), false)}.csv`);
     await FileUtil.writeTableToCSV(outputFilePath, reviewTable);
 }