+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Auto date for output no longer uses time, only dates
+- Review outputs now save the ENTIRE review database to file
 
 ## [0.3.1] - 2019-09-09
 ### Fixed